+++ include/cinolib/meshes/abstract_polyhedralmesh.h
@@ -122,10 +122,10 @@ class AbstractPolyhedralMesh : public AbstractMesh<M,V,E,P>
 
         //::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
 
-        virtual uint verts_per_poly(const uint pid) const override { return this->p2v.at(pid).size();   }
-        virtual uint faces_per_poly(const uint pid) const          { return this->polys.at(pid).size(); }
-        virtual uint verts_per_face(const uint fid) const          { return this->faces.at(fid).size(); }
-        virtual uint edges_per_face(const uint fid) const          { return this->faces.at(fid).size(); }
+        virtual uint verts_per_poly(const uint pid) const override { return uint(this->p2v.at(pid).size());   }
+        virtual uint faces_per_poly(const uint pid) const          { return uint(this->polys.at(pid).size()); }
+        virtual uint verts_per_face(const uint fid) const          { return uint(this->faces.at(fid).size()); }
+        virtual uint edges_per_face(const uint fid) const          { return uint(this->faces.at(fid).size()); }
 
         //::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
